fix: fix a bug that left the buffer imbalanced
  * fix a bug that was happening with bisecting spans, where the
    lack of any additional inserted spans was causing the S-Buffer
    to fail to balance itself, resulting in an unstable state
  * also add an assertion macro and use it in debug configuration
    to make sure the buffer is properly balanced at every step of
    the way

+++ build.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-gcc -c ./s_buffer.c -o ./s_buffer.o -fPIC -v && \
-gcc -shared ./s_buffer.o -o ./libsbuffer.so -v && \
-rm -rf ./s_buffer.o
+if [[ $1 == "-d" ]]; then
+    gcc -shared -DDEBUG ./s_buffer.c -o ./libsbuffer.so -fPIC -v -g
+else
+    gcc -c ./s_buffer.c -o ./s_buffer.o -fPIC -v && \
+    gcc -shared ./s_buffer.o -o ./libsbuffer.so -v && \
+    rm -rf ./s_buffer.o
+fi
